requirements.txt
pygame>=2.5.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.5.0
streamlit>=1.30.0
